refactor(layout): Move helpers to layout folder
