+++ src/app/layout/desktop-split/desktop-split.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { DataService } from '../data.service';
+import { DataService } from '../../data.service';
 
 @Component({
   selector: 'app-desktop-split',
