+++ src/app/layout/event-sidebar-desktop/event-sidebar-desktop.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input, OnChanges } from '@angular/core';
-import { DataService } from '../data.service';
+import { DataService } from '../../data.service';
 
 @Component({
   selector: 'app-event-sidebar-desktop',
