+++ src/app/layout/navmenu/navmenu.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { DataService } from '../data.service';
+import { DataService } from '../../data.service';
 import { Router } from '@angular/router';
-import { API } from '../../api';
+import { API } from '../../../api';
 
 const DEFAULT_USERNAME = 'Guest';
 const DEFAULT_LDAP = 'IITB User';
